# frontend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IoT Telehealth System API",
    "version": "1.0.0",
    "description": "Tài liệu API cho hệ thống theo dõi sức khỏe IoT"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT token nhận được sau khi đăng nhập"
      },
      "ConsentSessionHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-consent-session-token",
        "description": "Bắt buộc với bác sĩ khi truy cập dữ liệu theo consent session"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@gmail.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "doctor",
              "patient"
            ],
            "example": "patient"
          },
          "full_name": {
            "type": "string",
            "example": "Nguyễn Văn A"
          },
          "phone": {
            "type": "string",
            "example": "0901234567"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "is_verified": {
            "type": "boolean",
            "example": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 120
          },
          "pages": {
            "type": "integer",
            "example": 12
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "summary": "Danh sách người dùng (phân trang, tìm kiếm, lọc)",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Tìm kiếm theo email hoặc tên"
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "doctor",
                "patient"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "verified",
                "unverified"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách người dùng",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Danh sách người dùng"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Chưa xác thực"
          },
          "403": {
            "description": "Không có quyền"
          }
        }
      },
      "post": {
        "summary": "Tạo người dùng mới",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "role"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "doctor",
                      "patient"
                    ]
                  },
                  "full_name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "403": {
            "description": "Không có quyền"
          }
        }
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "summary": "Chi tiết người dùng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết người dùng"
          },
          "403": {
            "description": "Không có quyền"
          },
          "404": {
            "description": "Người dùng không tồn tại"
          }
        }
      },
      "put": {
        "summary": "Cập nhật thông tin người dùng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "full_name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "doctor",
                      "patient"
                    ]
                  },
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "403": {
            "description": "Không có quyền"
          }
        }
      },
      "delete": {
        "summary": "Vô hiệu hóa người dùng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vô hiệu hóa thành công"
          },
          "403": {
            "description": "Không có quyền"
          }
        }
      }
    },
    "/api/admin/users/{id}/role": {
      "patch": {
        "summary": "Thay đổi vai trò người dùng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "doctor",
                      "patient"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật vai trò thành công"
          },
          "403": {
            "description": "Không có quyền"
          }
        }
      }
    },
    "/api/admin/users/{id}/status": {
      "patch": {
        "summary": "Bật/tắt trạng thái người dùng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "is_active"
                ],
                "properties": {
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật trạng thái thành công"
          },
          "403": {
            "description": "Không có quyền"
          }
        }
      }
    },
    "/api/admin/users/{id}/reset-password": {
      "post": {
        "summary": "Đặt lại mật khẩu người dùng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mật khẩu tạm thời đã được tạo.\nNgười dùng phải đổi mật khẩu sau khi đăng nhập.\n",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "tempPassword": {
                      "type": "string",
                      "example": "a1b2c3d4e5f6"
                    },
                    "note": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền"
          },
          "404": {
            "description": "Người dùng không tồn tại"
          }
        }
      }
    }
  }
}